Create room (#53)
* Adds Django REST Framework

Adds Django REST Framework and Tag API

* Adding Room Get to REST API

This is identical to the Tag API that has already been added.

* Adding Question REST API

This is a dependancy of the Room REST API

* Adding iDjango REST Swagger

This enhances our web-browsable REST API. Swagger will provide an easy, interactive
spec of the API we are crafting.

Fixes #48

* Room create no longer requires code and questions

The 'code' and 'question' fields are marked as 'read_only' in the serializer, meaning they are ignored in the Create and Update operations.

* Adding supplementary data to Room Create API

Room create now accepts a list of tags, as well as the number of
questions. Additionally, the response (and related GET request) returns
room name and code, hiding other information

* Populating room with questions, based on requested tags and number of questions

A populate method has been added to the Room Model to generate a random set of
questions tagged with the specified tags, up to the number of questions requested.