Adding new contestant WebSocket message handling
This message will first verify that the request room exists, and if
it does it will add the contestant to the room. The conection is then
associated with the contestant, storing the room the contestant is a
member of within the session. Additionally the connection's reply channel
is added to the room's reply group, for use with all furure room broadcasts