Fix failing Contestant websocket unittest
The Contestant model was returning an un-ordered list (as is expected)
but the unittest was expecting a particular order.
I decided to make the model return an ordered dictionary so the test
passes. It's a pretty minor change and it's not a bad thing to enforce
some sort of order on the list that the server returns.

Fixes #91